[UPD] px -> percent
update movement and hitbox to have percentage insted of px

+++ ECS/Interface/Include/Components/sprite_data_component.hpp
@@ -22,7 +22,7 @@ template <typename T> struct Vector {
 struct SpriteDataComponent {
     SpritePath spritePath;
     Vector<uint32_t> offSet;
-    Vector<uint32_t> dimension;
+    Vector<float> dimension;
     Vector<float> scale;
     uint32_t type;
 };
